requirements.txt
asgiref==3.8.1
Django==5.0.6
django-appconf==1.0.6
django-filter==24.2
django-htmx==1.18.0
django-redis==5.4.0
django-select2==8.1.2
mysql-connector-python==8.4.0
mysqlclient==2.2.4
redis==5.0.7
setuptools==70.1.1
sqlparse==0.5.0
tzdata==2024.1
wheel==0.43.0
